fix: moving files to src to fix incorrect placement
